refactor(1794): S2 — move SkillRuntime → reyn.skill.skill_runtime (budget TYPE_CHECKING-gate)
#1794 skill-execution consolidation, stage 2 (per FP-0049). Moves the public
SkillRuntime entry from the loose top-level module into reyn.skill, with the
one reyn.runtime edge resolved by TYPE_CHECKING-gating (no inversion needed).

- git mv src/reyn/skill_runtime.py → src/reyn/skill/skill_runtime.py.
- The only reyn.runtime import — `from reyn.runtime.budget.budget import
  BudgetTracker` — is annotation-only (budget_tracker is held + forwarded, never
  called; used only in two `: BudgetTracker | None` hints, lazy via
  `from __future__ import annotations`). Moved under `if TYPE_CHECKING:` →
  reyn.skill takes ZERO runtime dependency on reyn.runtime (the layer invariant;
  the S1 layer gate verifies it).
- Repoint all 18 importers `reyn.skill_runtime` → `reyn.skill.skill_runtime`
  (incl. reyn/__init__'s lazy SkillRuntime re-export + reyn.skill.sub_skill_runner,
  now intra-package).

Behavior-preserving (not byte-identical — the annotation-gate edit). Gate:
straggler 0 (verify_package_move --new); layer gate PASS (reyn.skill runtime-
import-free); replay green; 478 pass across the importer areas (cli/session/
skill_runner/eval/run_id/replay). ruff clean.

(The one local swebench test failure is pre-existing env — swebench installed
locally makes the missing-honest-skip test take the real-eval path; verified
identical on clean base. CI without the eval extra honest-skips → passes.)

Refs #1794.
