+++ src/reyn/interfaces/cli/commands/run.py
@@ -134,7 +134,7 @@ def run(args: argparse.Namespace) -> None:
         from reyn.skill.skill_paths import is_stdlib_skill
         unsafe_python = is_stdlib_skill(loaded.skill_md.parent)
     from reyn.config import _find_project_root, load_project_context
-    from reyn.skill_runtime import SkillRuntime
+    from reyn.skill.skill_runtime import SkillRuntime
     from reyn.user_intervention import StdinInterventionBus
     project_root = _find_project_root(Path.cwd())
     project_context = load_project_context(session.config, project_root)
