+++ src/reyn/runtime/session.py
@@ -75,7 +75,7 @@ from reyn.security.permissions.permissions import PermissionResolver
 from reyn.services.compaction.engine import CompactionEngine
 from reyn.skill.skill_paths import SkillNotFoundError, resolve_skill_path, stdlib_root
 from reyn.skill.skill_registry import SkillRegistry
-from reyn.skill_runtime import SkillRuntime
+from reyn.skill.skill_runtime import SkillRuntime
 from reyn.user_intervention import (
     InterventionAnswer,
     InterventionChoice,
