+++ src/reyn/skill/sub_skill_runner.py
@@ -71,7 +71,7 @@ async def invoke_sub_skill(
     require approvals to be in place beforehand via reyn.yaml / reyn.local.yaml
     or .reyn/approvals.yaml (per documented permission model).
     """
-    from reyn.skill_runtime import SkillRuntime
+    from reyn.skill.skill_runtime import SkillRuntime
 
     agent = SkillRuntime(
         model=model,
